Create Child-Help Non Profit Org.
2. You are in charge or data collection for your team. You are tasked with collecting household
information on child ages, height and weight. If there are no children, then we need parent age
and education levels based on the provided criteria. Write a dynamic form that is able to
capture an unknown number of children capturing the required fields.
a) Submit your form in XFORM XLS format